requirements.txt
click @ file:///Users/ec2-user/ci_py311/click_1678318124837/work
Flask @ file:///Users/ec2-user/ci_py311/flask_1678380788105/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///Users/ec2-user/ci_py311/jinja2_1678317138271/work
MarkupSafe @ file:///Users/ec2-user/ci_py311/markupsafe_1678316973889/work
pyactiveresource==2.2.2
PyJWT==2.8.0
PyYAML==6.0.1
ShopifyAPI==12.3.0
six==1.16.0
Werkzeug @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_fegt3z60s3/croot/werkzeug_1679489762631/work
dynaconf==3.2.0
